Move executor back (broke /api/tasks/ endpoint)
