+++ pronto/__init__.py
@@ -1,11 +1,61 @@
-from datetime import timedelta
+from concurrent.futures import ThreadPoolExecutor
+from datetime import datetime, timedelta
 
 from flask import Flask, g, session
+
+
+class Executor(object):
+    def __init__(self):
+        self.executor = ThreadPoolExecutor()
+        self._tasks = {}
+
+    def submit(self, name, fn, *args, **kwargs):
+        self.update()
+        if name not in self._tasks:
+            self._tasks[name] = {
+                "name": name,
+                "future": self.executor.submit(fn, *args, **kwargs),
+                "time": datetime.now(),
+                "status": None
+            }
+
+    def has(self, name):
+        self.update()
+        return name in self._tasks
+
+    def update(self):
+        names = list(self._tasks)
+        for name in names:
+            task = self._tasks[name]
+            future = task["future"]
+            if future.done():
+                if future.exception() is not None:
+                    # Call raised: error
+                    task["status"] = False
+                elif (datetime.now() - task["time"]).total_seconds() > 3600:
+                    # Finished more than one hour ago: clean
+                    del self._tasks[name]
+                else:
+                    # Recently successfully finished
+                    task["status"] = True
+
+    @property
+    def tasks(self):
+        self.update()
+        tasks = []
+        for task in sorted(self._tasks.values(), key=lambda t: t["time"]):
+            tasks.append({
+                "name": task["name"],
+                "status": task["status"]
+            })
+
+        return tasks
 
 
 app = Flask(__name__)
 app.config.from_envvar("PRONTO_CONFIG")
 app.permanent_session_lifetime = timedelta(days=7)
+executor = Executor()
 
 
 def get_user():
