+++ pronto/views/api/entry/__init__.py
@@ -1,62 +1,8 @@
-from concurrent.futures import ThreadPoolExecutor
-from datetime import datetime
-
 from cx_Oracle import DatabaseError, STRING
 from flask import jsonify, request
 
-from pronto import app, db, get_user
+from pronto import app, db, executor, get_user
 from . import annotations, comments, go, references, relationships, signatures
-
-
-class Executor(object):
-    def __init__(self):
-        self.executor = ThreadPoolExecutor()
-        self._tasks = {}
-
-    def submit(self, name, fn, *args, **kwargs):
-        self.update()
-        if name not in self._tasks:
-            self._tasks[name] = {
-                "name": name,
-                "future": self.executor.submit(fn, *args, **kwargs),
-                "time": datetime.now(),
-                "status": None
-            }
-
-    def has(self, name):
-        self.update()
-        return name in self._tasks
-
-    def update(self):
-        names = list(self._tasks)
-        for name in names:
-            task = self._tasks[name]
-            future = task["future"]
-            if future.done():
-                if future.exception() is not None:
-                    # Call raised: error
-                    task["status"] = False
-                elif (datetime.now() - task["time"]).total_seconds() > 3600:
-                    # Finished more than one hour ago: clean
-                    del self._tasks[name]
-                else:
-                    # Recently successfully finished
-                    task["status"] = True
-
-    @property
-    def tasks(self):
-        self.update()
-        tasks = []
-        for task in sorted(self._tasks.values(), key=lambda t: t["time"]):
-            tasks.append({
-                "name": task["name"],
-                "status": task["status"]
-            })
-
-        return tasks
-
-
-executor = Executor()
 
 
 def _delete_entry(user, dsn, accession):
@@ -76,7 +22,6 @@ def _delete_entry(user, dsn, accession):
     finally:
         cur.close()
         con.close()
-        return datetime.now()
 
 
 @app.route("/api/entry/", methods=["PUT"])
